Feat : 유저 개인/전체 조회 & 회원 탈퇴 &  마이페이지 조회 ( #33 )

+++ src/main/java/com/example/eat_together/domain/cart/dto/CartItemResponseDto.java
@@ -16,7 +16,7 @@ public class CartItemResponseDto {
         this.itemId = cartItem.getId();
         this.menuName = cartItem.getMenu().getName();
         this.quantity = cartItem.getQuantity();
-        this.price = cartItem.getMenu().getPrice();
+        this.price = (int) cartItem.getMenu().getPrice(); // TODO : 이거 고침
         this.totalPrice = cartItem.getTotalPrice();
     }
 }
